requirements.txt
playwright==1.49.1
apscheduler==3.10.4
dnspython==2.7.0
pyyaml==6.0.2
jinja2==3.1.4
beautifulsoup4==4.12.3
resend==2.0.0
pytz==2024.1
